requirements_clean.txt
# Core dependencies
numpy>=1.24.0
pandas>=2.0.0
matplotlib>=3.7.0
seaborn>=0.12.0
tqdm>=4.65.0

# ML/AI dependencies
torch>=2.0.0
transformers>=4.35.0
tokenizers>=0.15.0
accelerate>=0.25.0

# API clients
openai>=1.0.0
anthropic>=0.15.0

# Dataset management
datasets>=2.14.0
huggingface_hub>=0.17.0
requests>=2.31.0

# Utilities
python-dotenv>=1.0.0
pydantic>=2.0.0
pytest>=7.4.0
black>=23.0.0
pyyaml>=6.0.0

# Optional for advanced features
scikit-learn>=1.3.0
scipy>=1.11.0
plotly>=5.17.0